% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_long_table.R
\name{topic_long_table}
\alias{topic_long_table}
\title{The eponymous wrapper for generating a topic long table.}
\usage{
topic_long_table(x, topic_cols, settings = c(), settings_after = c(),
  align_args = list(), caption_args = list(), header_args = list(),
  footer_args = list(), body_args = list())
}
\arguments{
\item{x}{The object to make into a topiclongtable (typically a data.frame or
matrix)}

\item{topic_cols}{Number of columns which are topics (where rowspans may be
used).}

\item{settings, settings_after}{Text to add before/after the table, typically
the output of \code{\link{topic_long_table_set}} and
\code{topic_long_table_set(reset=TRUE)}}

\item{align_args, caption_args, header_args, footer_args, body_args}{A list of
arguments to pass to \code{\link{topic_long_table_alignment}},
\code{\link{topic_long_table_caption}},
\code{\link{topic_long_table_header}},
\code{\link{topic_long_table_footer}}, \code{\link{topic_long_table_body}}}
}
\value{
A character scalar with class "topic_long_table" defining the LaTeX
  table (to be put into a LaTeX document)
}
\description{
The eponymous wrapper for generating a topic long table.
}
