% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_long_table_alignment.R
\name{topic_long_table_alignment}
\alias{topic_long_table_alignment}
\alias{topic_long_table_alignment.numeric}
\alias{topic_long_table_alignment.default}
\alias{topic_long_table_alignment.data.frame}
\alias{topic_long_table_alignment.matrix}
\title{Set alignment for a topiclongtable}
\usage{
topic_long_table_alignment(x, topic_cols, align = NULL,
  left_border = "|", inner_border = "|", right_border = "|", ...)

\method{topic_long_table_alignment}{numeric}(x, ...)

\method{topic_long_table_alignment}{default}(x, ...)

\method{topic_long_table_alignment}{data.frame}(x, topic_cols,
  align = NULL, left_border = "|", inner_border = "|",
  right_border = "|", ..., verbatim = NULL)

\method{topic_long_table_alignment}{matrix}(x, topic_cols, align = NULL,
  ...)
}
\arguments{
\item{x}{The object to make into a topiclongtable (typically a data.frame or
matrix)}

\item{topic_cols}{Number of columns which are topics (where rowspans may be
used).}

\item{align}{The alignment to use.  If \code{NULL}, defaults are used.}

\item{left_border, inner_border, right_border}{The left, inner, and right
borders of the table.}

\item{...}{Passed to other \code{topic_long_table_alignment} methods.}

\item{verbatim}{Use the provided character scalar verbatim without other
processing.}
}
\value{
A character scalar defining the alignment either for the current
  column (for individual classes) or the entire table (for table classes like
  data.frame and matrix).
}
\description{
Set alignment for a topiclongtable
}
\section{Methods (by class)}{
\itemize{
\item \code{numeric}: For numbers, align on decimal.

\item \code{default}: For anything not specified, left justify.

\item \code{data.frame}: For data.frames, alignment occurs by column class.

\item \code{matrix}: For matrices, alignment occurs by class.
}}

