% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_long_table.R
\name{topic_long_table_document_headers}
\alias{topic_long_table_document_headers}
\title{Assist the user to know which packages to load.}
\usage{
topic_long_table_document_headers()
}
\description{
Assist the user to know which packages to load.
}
