% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_long_table_header_footer.R
\name{topic_long_table_header}
\alias{topic_long_table_header}
\alias{topic_long_table_footer}
\title{Generate the table header or footer}
\usage{
topic_long_table_header(
  x,
  col_names = NULL,
  above_col_names = "\\\\hline",
  below_col_names = "\\\\hline",
  subsequent_page_notification = "\\\\ldots continued",
  latex_header = NULL,
  verbatim = NULL
)

topic_long_table_footer(
  x,
  bottom_border = "\\\\hline",
  bottom_all_pages = NULL,
  bottom_last_page = NULL,
  subsequent_page_notification = "continued \\\\ldots",
  verbatim = NULL
)
}
\arguments{
\item{x}{The object to generate the header for.}

\item{col_names}{The column names to use for the table header (defaults to
\code{Hmisc::latexTranslate(colnames(x))} if \code{NULL}).  This must have
the same length as \code{ncol(x)}.}

\item{above_col_names, below_col_names}{LaTeX text to place above/below the
col_names text (typically this will be a horizontal line with "\\hline").}

\item{subsequent_page_notification}{LaTeX text to place above the subsequent
page headers}

\item{latex_header}{A fully-defined LaTeX header to use in place of the
auto-generated header.  (If a character scalar, it is returned as-is; if
character a vector, it is joined with " \\\\\\n" separating the elements.)}

\item{verbatim}{Use the provided character scalar verbatim without other
processing.}

\item{bottom_border}{The border below the tabular text on all pages.}

\item{bottom_all_pages, bottom_last_page}{LaTeX text below the bottom border
on all (including the last) or only the last page, respectively.}
}
\value{
A character string of table header for a topic_long_table (a single
  string, not a vector).
}
\description{
Generate the table header or footer
}
\section{Functions}{
\itemize{
\item \code{topic_long_table_footer}: Generate the footer.
}}

