% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_long_table_set.R
\name{topic_long_table_set}
\alias{topic_long_table_set}
\title{Set values that control marking of continued cells and positioning within
continued cells.}
\usage{
topic_long_table_set(continuation_code = NULL, vpos = NULL,
  reset = FALSE)
}
\arguments{
\item{continuation_code}{LaTeX text to add after a continued cell value.}

\item{vpos}{Vertical position to use within topic-combined cells (valid
values are given by providing an invalid value like "")}

\item{reset}{Provide code to reset to defaults}
}
\value{
A character scalar or vector of settings
}
\description{
Set values that control marking of continued cells and positioning within
continued cells.
}
