% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_long_table_body.R
\name{topic_long_table_body}
\alias{topic_long_table_body}
\title{Generate the body of a topiclongtable}
\usage{
topic_long_table_body(
  x,
  topic_cols,
  row_border = "\\\\TopicLine",
  latex_clean = TRUE,
  verbatim = NULL
)
}
\arguments{
\item{x}{The object to make into a topiclongtable (typically a data.frame or
matrix)}

\item{topic_cols}{Number of columns which are topics (where rowspans may be
used).}

\item{row_border}{The LaTeX command between rows}

\item{latex_clean}{Should cleaning occur?  Cleaning is: If a data.frame or
matrix, the contents are run through
\code{Hmisc::latexTranslate(format(x))} prior to generating the table body.}

\item{verbatim}{Use the provided character scalar verbatim without other
processing.}
}
\value{
A character string for use in a topic_long_table body.
}
\description{
Generate the body of a topiclongtable
}
