% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_long_table_caption.R
\name{topic_long_table_caption}
\alias{topic_long_table_caption}
\title{Generate a LaTeX table caption}
\usage{
topic_long_table_caption(
  x,
  text = NULL,
  label = NULL,
  combine_short_long = TRUE,
  latex_clean = TRUE,
  verbatim = NULL
)
}
\arguments{
\item{x}{The object to make into a topiclongtable (typically a data.frame or
matrix)}

\item{text}{Text to place in the caption.  See Details.}

\item{label}{The label for the caption}

\item{combine_short_long}{Should the short and long captions be combined for
the long caption?  See Details.}

\item{latex_clean}{Should cleaning occur?  Cleaning is: If a data.frame or
matrix, the contents are run through
\code{Hmisc::latexTranslate(format(x))} prior to generating the table body.}

\item{verbatim}{Use the provided character scalar verbatim without other
processing.}
}
\value{
The caption for inclusion in a table or an empty string if no caption
  is provided.
}
\description{
Generate a LaTeX table caption
}
\details{
For \code{text}, if \code{NULL}, the caption attribute for \code{x}
  will be used, if it exists.  If neither is provided, the caption will be an
  empty string (i.e. no caption).  If the length of the character vector for
  the caption text (as defined earlier in this paragraph) is 1, the caption
  will be used as-is; if the length is 2 and \code{combine_short_long ==
  TRUE}, then the first element will be used as the short caption and the
  first and second will be pasted together for the long caption.
}
