% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic_long_table_caption.R
\name{topic_long_table_caption}
\alias{topic_long_table_caption}
\title{Generate a LaTeX table caption}
\usage{
topic_long_table_caption(x, text = NULL, label = NULL)
}
\arguments{
\item{x}{The object to make into a topiclongtable (typically a data.frame or
matrix)}

\item{text}{Text to place in the caption.  If \code{NULL}, the caption
attribute for \code{x} will be used, if it exists.  If neither is provided,
the caption will be an empty string.}

\item{label}{The label for the caption}
}
\value{
The caption for inclusion in a table or an empty string if no caption
  is provided.
}
\description{
Generate a LaTeX table caption
}
